US State Facts

Updated version of the datasets::state.x77 matrix, which provides eights statistics from the 1970's. This version is a modern data frame format with updated (and alternative) statistics.

facts

Format

A tibble with 52 rows and 9 variables:

name

Full state name

population

Population estimate (September 26, 2019)

votes

Votes in the Electoral College (following the 2010 Census)

admission

The data which the state was admitted to the union

income

Per capita income (2018)

life_exp

Life expectancy in years (2017-18)

murder

Murder rate per 100,000 population (2018)

college

Percent adult population with at least a bachelor's degree or greater (2019)

heat

Mean number of degree days (temperature requires heating) per year from 1981-2010

Source

Population: https://www2.census.gov/programs-surveys/popest/datasets/2010-2018/state/detail/SCPRC-EST2018-18+POP-RES.csv

Electoral College: https://www.archives.gov/electoral-college/allocation

Income: (Moved Census table ACSST1Y2018.S1903 on income)

GDP: (Moved BEA dataset on GDP)

Literacy: https://nces.ed.gov/naal/estimates/StateEstimates.aspx

Life Expectancy: https://web.archive.org/web/20231129160338/https://usa.mortality.org/

Murder: https://ucr.fbi.gov/crime-in-the-u.s/2018/crime-in-the-u.s.-2018/tables/table-4/table-4.xls/output.xls

Education: (Noved Census table S1501 on education)

Temperature: (Moved NOAA dataset on temperature)